pyproject.toml
[project]
name = "facilitator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.13",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "websockets>=15.0.1",
]
